My project/hackson/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
jinja2==3.1.6

# AI services
openai==1.57.0
anthropic==0.39.0
dashscope==1.20.0

# File parsing
PyPDF2==3.0.1
python-docx==1.1.2
markdown==3.7

# Utilities
pydantic==2.9.0
aiofiles==24.1.0
python-dotenv==1.0.1
